# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Travelode API
host: 127.0.0.1:28825
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
...
paths:
  /user:
    # binds a127 app logic to a route
    x-swagger-router-controller: user

    get:
      tags:
        - User
      summary: Returns users list
      operationId: findUsers
      parameters:
        - name: page
          in: query
          description: The page number to get. Default to 0.
          required: false
          type: string
        - name: size
          in: query
          description: The size of the page to get. Default to 50.
          required: false
          type: string
        - name: email
          in: query
          description: Find by email id
          required: false
          type: string
        - name: fullname
          in: query
          description: Find by full name
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserListResponse"
        # responses may fall through to errors
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    post:
      tags:
        - User
      summary: Create new user
      operationId: createUser
      consumes:
        - multipart/form-data
      produces:
          - application/json
      parameters:
        - in: formData
          name: email
          description: Email id of the user, has to be unique
          required: true
          type: string
          format: email
        - in: formData
          name: password
          description: Password of the user
          required: true
          type: string
          format: password
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/CreatedResponse"
        # responses may fall through to errors
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/{id}:
    x-swagger-router-controller: user

    get:
      tags:
        - User
      summary: Returns a single user
      operationId: getUserById
      parameters:
        - name: id
          in: path
          description: The User Id
          required: true
          type: integer
          minimum: 1
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserListResponse"
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      tags:
        - User
      summary: Update a user
      operationId: updateUserById
      parameters:
        - name: id
          in: path
          description: The user id to update
          required: true
          type: integer
          minimum: 1
        - name: userData
          in: body
          description: The user data to update
          required: true
          schema:
            type: object
            properties:
              fullname:
                type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      tags:
        - User
      summary: Delete a user
      operationId: deleteUserById
      parameters:
        - name: id
          in: path
          description: The user id to delete
          required: true
          type: integer
          minimum: 1
      responses:
        "204":
          description: User is removed
        "404":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /user/login:
    # binds a127 app logic to a route
    x-swagger-router-controller: user

    post:
      tags:
        - User
      summary: Retrieve user by login credentials
      operationId: retrieveUserByLogin
      consumes:
        - multipart/form-data
      produces:
          - application/json
      parameters:
        - in: formData
          name: email
          description: Email id of the user
          required: true
          type: string
          format: email
        - in: formData
          name: password
          description: Password of the user
          required: true
          type: string
          format: password
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserListResponse"
        "401":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  CreatedResponse:
    required:
      - message
      - id
    properties:
      message:
        type: string
      id:
        type: number
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  User:
    type: object
    properties:
      id:
        type: number
        description: User Id
      email:
        type: string
        description: User Email
      fullname:
        type: string
        description: User Full Name
      photo:
        description: User Photo URL
        type: string
      created:
        description: User Creation Date Time
        type: string
        format: date-time
      updated:
        description: User Update Date time
        type: string
        format: date-time
  UserListResponse:
    required:
      - users
    properties:
      # The array of users
      users:
        type: array
        items:
          $ref: "#/definitions/User"
...